reduce code duplication between contributors and abstract agents subjects
